% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{classifier}
\alias{classifier}
\title{Classify function}
\usage{
classifier(tweets, pos.words, neg.words, keyword)
}
\arguments{
\item{tweets}{is the dataframe of tweets. pos.words and neg.words come from the lexicon (see example). The keyword is the topic used to subset the data}
}
\description{
This function is based on Pablo Barbera's workshop in Internet data collection. You need to load the lexicon first and create the pos.words and neg.words objects first (see example)
}
\examples{
lexicon <- read.csv("https://raw.githubusercontent.com/pablobarbera/workshop/master/lexicon.csv", stringsAsFactors = F)
pos.words <- lexicon$word[lexicon$polarity == "positive"]
neg.words <- lexicon$word[lexicon$polarity == "negative"]
classifier(tweets, pos.words, neg.words, keyword = "")
}
\author{
Javier Sajuria
}
\keyword{Twitter,}
\keyword{subset}

