% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{classify}
\alias{classify}
\title{Classify function}
\usage{
classify(words, pos.words, neg.words)
}
\arguments{
\item{words}{is the text, pos.words and neg.words come from the lexicon. Use the example first}
}
\description{
This function is based on Pablo Barbera's workshop in Internet data collection
}
\examples{
lexicon <- read.csv("https://raw.githubusercontent.com/pablobarbera/workshop/master/lexicon.csv", stringsAsFactors = F)
pos.words <- lexicon$word[lexicon$polarity == "positive"]
neg.words <- lexicon$word[lexicon$polarity == "negative"]
clean.tweets()
}
\author{
Javier Sajuria
}
\keyword{Twitter,}
\keyword{subset}

