% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{clean.tweets}
\alias{clean.tweets}
\title{Clean Tweets function}
\usage{
clean.tweets(text)
}
\arguments{
\item{text}{is the text column of the tweets data frame}
}
\description{
This function is based on Pablo Barbera's workshop in Internet data collection
}
\examples{
clean.tweets()
}
\author{
Javier Sajuria
}
\keyword{Twitter,}
\keyword{subset}

