% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{subset_cons}
\alias{subset_cons}
\title{Subset Conservative function}
\usage{
subset_cons(df)
}
\arguments{
\item{df}{is the dataframe of tweets. The text of the tweets whould be under a column named "text"}
}
\description{
This function creates a subset from the tweets dataframe (collected and parsed using ''streamR'') that corresponds to a group of political terms related to Conservative Party
}
\examples{
subset_cons()
}
\author{
Javier Sajuria
}
\keyword{Twitter,}
\keyword{subset}

