% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{subset_polis}
\alias{subset_polis}
\title{Subset polis function}
\usage{
subset_polis(df)
}
\arguments{
\item{df}{is the dataframe of tweets. The text of the tweets whould be under a column named "text"}
}
\description{
This function creates a subset from the tweets dataframe (collected and parsed using ''streamR'') that corresponds to a group of political terms:
'europe','ukip','farage','labour','miliband','conservative','cameron','libdem','clegg','lucas','green party','tory','politic','election','EU','parliament','government','pmq','vote','opposition'
}
\examples{
subset_polis()
}
\keyword{Twitter,}
\keyword{subset}

