% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{subset_ukip}
\alias{subset_ukip}
\title{Subset UKIP function}
\usage{
subset_ukip(df)
}
\arguments{
\item{df}{is the dataframe of tweets. The text of the tweets whould be under a column named "text"}
}
\description{
This function creates a subset from the tweets dataframe (collected and parsed using ''streamR'') that corresponds to a group of political terms related to UKIP
}
\examples{
subset_ukip()
}
\author{
Javier Sajuria
}
\keyword{Twitter,}
\keyword{subset}

